st_ai/web_/flask-web-app/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
Jinja2==3.1.2
Werkzeug==2.2.2
sentence-transformers
faiss-cpu
nltk
requests
pymupdf
pillow